feat(decoder): v0 new token vaults and pool params

+++ decoders/omnipair_decoder/types/user_position.rs
@@ -11,7 +11,7 @@ pub struct UserPosition {
     pub collateral1_applied_min_cf_bps: u16,
     pub collateral0: u64,
     pub collateral1: u64,
-    pub debt0_shares: u64,
-    pub debt1_shares: u64,
+    pub debt0_shares: u128,
+    pub debt1_shares: u128,
     pub bump: u8,
 }
